Fix the neuron resampler approach (#141)
This was normalising across an incorrect dimension and not firing at the correct point.

+++ sparse_autoencoder/metrics/train/neuron_activity_metric.py
@@ -134,7 +134,7 @@ class NeuronActivityHorizonData:
             thresholds: Thresholds for almost dead neurons.
         """
         self._steps_since_last_calculated = 0
-        self._neuron_activity = torch.zeros(number_learned_features, dtype=torch.int64)
+        self._neuron_activity = torch.zeros(number_learned_features, dtype=torch.int)
         self._thresholds = thresholds
 
         # Get a precise activation_horizon
